requirements.txt
numpy
pandas
scikit-learn
matplotlib
websocket-client
requests
streamlit
transformers
torch
pyyaml
streamlit
